refatora o serviço de API para integrar essas funções utilitarias e adiciona documentacao para diversas funções
